blog/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
boto3==1.26.59
botocore==1.29.165
celery==5.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.1
django-cors-headers==4.0.0
django-storages==1.11.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
install==1.3.5
jmespath==1.0.1
kombu==5.3.1
mysqlclient==2.2.0
Pillow==9.1.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
redis==5.0.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
whitenoise==6.2.0
gunicorn==20.1.0
